backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
